refactor(agent): split main.rs into startup and shutdown modules
startup.rs owns the ladder this branch introduced — bootstrap/Core, the
dormancy gate, the arming-time state watchers, the IPC server glue —
and shutdown.rs the signal listeners plus the hook-releasing exit path.
Also folds the bootstrap result straight into a let-else pattern and
lets the select loop read the watcher struct's fields directly instead
of destructuring them into renamed locals.
